playground/social_stuff/go_neo4j/requirements.txt
aiofiles==0.8.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.13
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
confluent-kafka==1.8.2
cryptography==3.3.2
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
ecdsa==0.18.0
elasticsearch==7.7.0
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.81.0
fastjsonschema==2.16.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.9.1
ipython==7.32.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==0.14.0
jsonschema==4.16.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
minio==7.1.6
mistune==2.0.4
mlaas-tools2==1.2.0
mlaas2-connection==0.0.1
nbclient==0.6.8
nbconvert==7.2.1
nbformat==5.6.0
neo4j==4.4.3
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.12
numpy==1.19.1
openpyxl==3.0.10
packaging==21.3
pandas==1.1.2
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pyarrow==9.0.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.2
Pygments==2.12.0
pylint==2.15.6
PyNaCl==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
redis==3.0.1
requests==2.27.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
terminado==0.16.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typed-ast==1.5.4
typing_extensions==4.0.0
urllib3==1.26.12
uvicorn==0.13.0
uvloop==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
wrapt==1.14.1
yamllint==1.26.3
zipp==3.8.1
